Mortadelo
Hi Federico, I've been reading about Mortadelo and tried to run it on
my FC6 today. If found just one big issue:

FC6 comes with option 'requiretty' on /etc/sudoers, so when Mortadelo
run stap (without a tty) the sudo that stap run returns claiming about
no tty available. On FC6 it's just a matter of remove 'requiretty'
option (which is Off on default sudo).

I was just wondering if this happens on another distros and if
Mortadelo can run stap on a shell to get an tty to satisfy sudo's
desires when using requiretty option.

The strange fact it's that stap runs sudo even when I'm running stap
as root! btw, I don't know if this is an issue of stap.

Best regards,

PS. I think that we have some common friends here in Brazil, Izabel
Valverde from Curitiba, I was a GNOME SoC student last year and was at
GUADEC last year, unfortunatelly I couln'd have a chance to meet you
personally.

PS2: Follow a patch to add this note on know issues section of Mortadelo README.

-- Paulo Henrique